### 2024-04-16 #6631 @editorialbot
[PRE REVIEW]: Pycashier: cash in on DNA barcode tags
**Submitting author:** @daylinmorgan (<a href="http://orcid.org/0000-0002-4218-2805">Daylin Morgan</a>)
**Repository:** https://github.com/brocklab/pycashier
**Branch with paper.md** (empty if default branch): joss
**Version:** v2024.1004
