### 2024-04-28 #6631 @Kevin-Mattheus-Moerman
Dear author, @daylinmorgan thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues (it is okay if those highlighted really do not have a DOI). You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
- [ ] The README is very short in general. Typically has clear description (usually not overly technical), installation instructions, as well as things like a getting started section, links to documentation (I think you have this but a small section could also work), links to testing, links to the license, links to the contributing guidelines. So please extend the README to include these.
- [ ] In your affiliation in the paper, please spell out USA as United States of America.
