### 2024-05-01 #6631 @daylinmorgan
Thank you for the helpful comments!

As requested I've added a [`CONTRIBUTING.md` and expanded the `README.md`](https://github.com/brocklab/pycashier/pull/26). And [spelled out USA](https://github.com/brocklab/pycashier/commit/055f8e6c06b01525f15f80842aeff2d5ae51993f) within the affiliation in the paper.

I'm unable to provide DOI's for the two missing since both are citations directly to URL's that have no accompanying DOI's.

Let me know anything else that should be updated.
