+++ paper/paper.bib
@@ -29,20 +29,23 @@
   abstract = {Quality control and preprocessing of FASTQ files are essential to providing clean data for downstream analysis. Traditionally, a different tool is used for each operation, such as quality control, adapter trimming and quality filtering. These tools are often insufficiently fast as most are developed using high-level programming languages (e.g. Python and Java) and provide limited multi-threading support. Reading and loading data multiple times also renders preprocessing slow and I/O inefficient.We developed fastp as an ultra-fast FASTQ preprocessor with useful quality control and data-filtering features. It can perform quality control, adapter trimming, quality filtering, per-read quality pruning and many other operations with a single scan of the FASTQ data. This tool is developed in C++ and has multi-threading support. Based on our evaluation, fastp is 2--5 times faster than other FASTQ preprocessing tools such as Trimmomatic or Cutadapt despite performing far more operations than similar tools.The open-source code and corresponding instructions are available at https://github.com/OpenGene/fastp.}
 }
 
-@article{dhimolea2021,
-  title = {An {{Embryonic Diapause-like Adaptation}} with {{Suppressed Myc Activity Enables Tumor Treatment Persistence}}},
-  author = {Dhimolea, Eugen and {de Matos Simoes}, Ricardo and Kansara, Dhvanir and Al'Khafaji, Aziz and Bouyssou, Juliette and Weng, Xiang and Sharma, Shruti and Raja, Joseline and Awate, Pallavi and Shirasaki, Ryosuke and Tang, Huihui and Glassner, Brian J. and Liu, Zhiyi and Gao, Dong and Bryan, Jordan and Bender, Samantha and Roth, Jennifer and Scheffer, Michal and Jeselsohn, Rinath and Gray, Nathanael S. and Georgakoudi, Irene and Vazquez, Francisca and Tsherniak, Aviad and Chen, Yu and Welm, Alana and Duy, Cihangir and Melnick, Ari and Bartholdy, Boris and Brown, Myles and Culhane, Aedin C. and Mitsiades, Constantine S.},
+@article{emert2021,
+  title = {Variability within Rare Cell States Enables Multiple Paths toward Drug Resistance},
+  author = {Emert, Benjamin L. and Cote, Christopher J. and Torre, Eduardo A. and Dardani, Ian P. and Jiang, Connie L. and Jain, Naveen and Shaffer, Sydney M. and Raj, Arjun},
   year = {2021},
-  month = feb,
-  journal = {Cancer Cell},
+  month = jul,
+  journal = {Nature Biotechnology},
   volume = {39},
-  number = {2},
-  pages = {240-256.e11},
-  issn = {1535-6108},
-  doi = {10.1016/j.ccell.2020.12.002},
-  urldate = {2024-04-05},
-  abstract = {Treatment-persistent residual tumors impede curative cancer therapy. To understand this cancer cell state we generated models of treatment persistence that simulate the residual tumors. We observe that treatment-persistent tumor cells in organoids, xenografts, and cancer patients adopt a distinct and reversible transcriptional program resembling that of embryonic diapause, a dormant stage of suspended development triggered by stress and associated with suppressed Myc activity and overall biosynthesis. In cancer cells, depleting Myc or inhibiting Brd4, a Myc transcriptional co-activator, attenuates drug cytotoxicity through a dormant diapause-like adaptation with reduced apoptotic priming. Conversely, inducible Myc upregulation enhances acute chemotherapeutic activity. Maintaining residual cells in dormancy after chemotherapy by inhibiting Myc activity or interfering with the diapause-like adaptation by inhibiting cyclin-dependent kinase 9 represent potential therapeutic strategies against chemotherapy-persistent tumor cells. Our study demonstrates that cancer co-opts a mechanism similar to diapause with adaptive inactivation of Myc to persist during treatment.},
[…]
