### 2024-06-10 #6631 @lpantano
@Kevin-Mattheus-Moerman do you know why the bot didn't recognize the command? According to the docs, it seems fine.
