### 2024-06-14 #6891 @editorialbot
[REVIEW]: Pycashier: cash in on DNA barcode tags
**Submitting author:** @daylinmorgan (<a href="http://orcid.org/0000-0002-4218-2805">Daylin Morgan</a>)
**Repository:** https://github.com/brocklab/pycashier
**Branch with paper.md** (empty if default branch): joss
**Version:** v2024.1006
