### 2024-08-11 #6891 @LTLA
I didn't get a checklist so I just assumed I wasn't needed here anymore. Nonetheless I spent a little bit of time looking at the content so I can give a few preliminary thoughts.

The first is that the paper or the documentation should be clearer about the barcode designs that they support. Is the sequence of the variable region a random k-mer or is it drawn from some whitelist (e.g., 10X)? Is the user responsible for specifying the sequence of the constant flanking regions, or are they commonly used sequences that can be auto-detected from the data (e.g., Illumina adapters)? I don't see any options for specifying sequences in the documentation, either in the command line flags or in the configuration file parameters.

Authors should also consider providing some worked examples that a prospective user can run on their machines, otherwise it's hard to tell whether it meets their needs. (Certainly it would be hard for me to review it, given that I don't understand the experimental design enough to find appropriate test data.) This is relatively straightforward, just grab some FASTQs from a public dataset and run it through all the relevant Pycashier commands. Maybe you'll need to downsample it if the public data is too big, so that the workflow has a reasonable runtime. 10X Genomics also provides various CellRanger outputs on their website so the `scrna` workflow doesn't even need to start from the FASTQs.

The [CLI reference](https://docs.brocklab.com/pycashier/cli.html) shows images of the `--help` output of each command. This is quite annoying as a reader cannot easily copy/paste the flags.

Code looks fine from a cursory inspection. The only thing that comes to mind is that the system calls are done with hard-coded executable names, e.g., [`fastp`](https://github.com/brocklab/pycashier/blob/5d9965043ad56e814bc3334b339e8842736cc2cb/src/pycashier/sample.py#L133), [`cutadapt`](https://github.com/brocklab/pycashier/blob/5d9965043ad56e814bc3334b339e8842736cc2cb/src/pycashier/sample.py#L169). These commands might not be on the `PATH` for all users, or they might be deliberately renamed  (e.g., to support multiple versions of the command on the same system). So it would be helpful to allow users to set these commands via the CLI.
