### 2024-08-13 #6891 @lpantano
Thank you @LTLA! , you can generate your checklist with this command: `@editorialbot generate my checklist`
