### 2024-08-19 #6891 @adrientaudiere
Hi @daylinmorgan,

I started review pycashier. I appreciate the time you've spent making the tools available to other researchers. The documentation site is well-structured and useful. I have some first questions/remarks.

- Using a personal fastq.files I obtained an error "failed to convert fastq to tsv". Can you provide several example files to try the program? In any case, this will be useful to Pycashier users to see what is a typical file for Pycashier input. I finally founded one example file in the test/data/references folder of the github repositories but it’s a bit hidden. In link to other questions hereafter about specificity of cell metabarcoding, I wonder what are the special structures of fastq files needed for Pycashier.

In the same idea, an example usage is required by JOSS, at least in the documentation. I didn’t find one.

- Can you develop more the summary for people unfamiliar with the concept of barcode? Especially in the concept of cell lineage. I personally use metabarcoding at the ecological scales (environmental DNA) and when I read your paper, I sometimes forget the scope of your software (tracing of cell lineages). I know there is a trade-off between repetition and clarity, but I think you may underline even more the cell scope of your approach. For example, I found Holze et al. (2024) very clear about that point.

- In the statement of need, it will be useful to add the features that lack Pycashier compared to BARtab (tab 1 in Holze et al. 2024; e.g., No reference-based quantification and no sample parallelization in Pycashier) and NextClone to help reader choose the good software for their use. I note that parallelization seems available through the parameter --threads, contrary to what Holze and colleagues stated.

- A rapid illustration of users cases and Pycashier command would be useful. I made a basic scheme from what I understand from documentation.

Paired-end -> yes -> pycashier merge
|
-> no -> from scRNA-seq -> yes -> pycashier scrna  -> pychasier receipt
                |
                -> no -> pychasier extract -> pychasier receipt
