### 2024-08-26 4743fc6
docs(#29): add plaintext version of cli reference
