### 2024-08-26 a2197ab
docs(#29): add plaintext version of cli reference
