+++ docs/cli.md
@@ -1,21 +1,0 @@
-# CLI Reference
-
-## pycashier
-
-![pycashier](svgs/pycashier.svg)
-
-## pycashier receipt
-
-![pycashier receipt](svgs/pycashier-receipt.svg)
-
-## pycashier extract
-
-![pycashier extract](svgs/pycashier-extract.svg)
-
-## pycashier merge
-
-![pycashier merge](svgs/pycashier-merge.svg)
-
-## pycashier scrna
-
-![pycashier scrna](svgs/pycashier-scrna.svg)
