### 2024-08-26 38b0645
build: drop pdm for pixi only dev environment
