+++ pyproject.toml
@@ -3,11 +3,11 @@ name = "pycashier"
 description = "cash in on expressed barcode tags"
 authors = [{ name = "Daylin Morgan", email = "daylinmorgan@gmail.com" }]
 dependencies = [
-    "rich>=12.0.0",
-    "tomlkit>=0.10.1",
-    "click-rich-help>=22.1.0",
-    "click>=8.1.0",
-    "polars>=0.20.4",
+  "rich>=12.0.0",
+  "tomlkit>=0.10.1",
+  "click-rich-help>=22.1.0",
+  "click>=8.1.0",
+  "polars>=1.0.0",
 ]
 requires-python = ">=3.8"
 readme = "README.md"
@@ -22,8 +22,10 @@ documentation = "https://docs.brocklab.com/pycashier"
 pycashier = "pycashier.cli:main"
 
 [build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
+requires = ["setuptools>=64", "setuptools_scm>=8"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools_scm]
 
 [tool.mypy]
 check_untyped_defs = true
@@ -40,58 +42,33 @@ ignore_missing_imports = true
 
 [tool.ruff]
 select = [
-    # Pyflakes
-    "F",
-    # Pycodestyle
-    "E",
-    "W",
-    # isort
-    "I001"
+  # Pyflakes
+  "F",
+  # Pycodestyle
+  "E",
+  "W",
+  # isort
+  "I001",
 ]
 # docstrings are longer because of formatting
 # ruff will handle the code formatting for us
 ignore = ["E501"]
 
 [tool.coverage.run]
-omit = [
-    "tests/",
-    "env/"
-    ]
-
-
-[tool.pdm]
-version = { source = "scm" }
-# need python 3.9 for these which I usually have anyways
-ignore_package_warnings = ["sphinx*", "myst-parser"]
-
-[tool.pdm.dev-dependencies]
-dev = [
-    "pre-commit<3.0.0,>=2.15.0",
-    "mypy>=0.982",
-    "pytest>=7.2.0",
-    "pytest-cov>=4.1.0",
-]
-docs = [
-    "sphinx",
-    "sphinx-autobuild",
-    "sphinx-copybutton",
-    "myst-parser",
-    "shibuya",
-]
-
+omit = ["tests/", "env/"]
 
 [tool.pixi.project]
 name = "pycashier"
 authors = ["Daylin Morgan <daylinmorgan@gmail.com>"]
-channels = ["conda-forge" , "bioconda"]
+channels = ["conda-forge", "bioconda"]
 platforms = ["linux-64"]
 
 
 [tool.pixi.environments]
-dev = ["base", "bioconda", "dev", "build"]
-prod = ["base","bioconda", "prod"]
-build = ["build"]
-test = ["bioconda", "dev", "test"]
+dev = ["base", "bioconda", "dev", "build", "docs", "tutorial"]
+prod = ["base", "bioconda", "prod"]
+build = ["base", "build"]
+test = ["base", "bioconda", "dev", "test"]
 test38 = ["bioconda", "dev", "test", "test38"]
 test39 = ["bioconda", "dev", "test", "test39"]
 test310 = ["bioconda", "dev", "test", "test310"]
@@ -100,18 +77,32 @@ test310 = ["bioconda", "dev", "test", "test310"]
 pip = "*"
 
 [tool.pixi.feature.base.dependencies]
-# https://bioconda.github.io/faqs.html#what-versions-are-supported
[…]
