+++ docker/Dockerfile
@@ -1,9 +1,9 @@
-FROM ghcr.io/prefix-dev/pixi:0.18.0 AS build
+FROM ghcr.io/prefix-dev/pixi:0.28.1 AS build
 
 COPY . /app
 WORKDIR /app
-RUN pixi install --frozen --environment build
-RUN pixi install --frozen --environment prod
+RUN pixi install -vv --frozen --environment build
+RUN pixi install -vv --frozen --environment prod
 RUN pixi run -e build -vv build-wheel
 RUN pixi run -e prod -vv postinstall-production
 RUN pixi shell-hook -e prod > /shell-hook
@@ -12,7 +12,7 @@ RUN echo 'pycashier "$@"' >> /shell-hook
 # cleanup unnecessary files from environment to reduce image size
 RUN bash docker/purge.sh
 
-FROM debian:bookworm-slim as production
+FROM debian:bookworm-slim AS production
 
 RUN groupadd -g 2009 pycashier && useradd -u 2009 -g 2009 pycashier
 USER pycashier
