### 2024-08-26 4bb2d4f
build: drop pdm for pixi only dev environment
