### 2024-08-26 f5b5ca9
build: drop pdm for pixi only dev environment
