+++ .github/workflows/tests.yml
@@ -27,15 +27,12 @@ jobs:
           # - "test312"
     steps:
       - uses: actions/checkout@v4
-      - uses: prefix-dev/setup-pixi@v0.5.1
+      - uses: prefix-dev/setup-pixi@v0.8.1
         with:
-          pixi-version: v0.18.0
-          # setup-pixi tries to cache using a non-existent pyproject.lock
-          # cache: true
+          pixi-version: v0.28.2
+          cache: true
           manifest-path: pyproject.toml
-
-      - name: Install pycashier for tests
-        run: pixi run -e ${{ matrix.environment }} pdm install -d -G dev
+          environments: ${{ matrix.environment }}
 
       - name: Run tests
         run: pixi run -e ${{ matrix.environment }} pytest
