+++ paper/paper.md
@@ -22,48 +22,62 @@ bibliography: paper.bib
 
 # Summary
 
-`Pycashier` is a tool designed to extract DNA barcode tags from next generation sequencing data.
-It was originally developed for use with the
-ClonMapper Barcoding System [@alkhafaji2018;@gardner2024;@gardner2022].
+`Pycashier` is a tool designed to extract cellular DNA barcode tags from next generation sequencing data.
+These DNA barcode tags are heritable and stably integrated genetic markers useful for clonal tracking (CITE) (clontracr?)
+and lineage tracing (development) (gestalt) within *in vitro* and *in vivo* cell-based disease models.
+These exogenous cell-based DNA barcodes, when amplified from genomic DNA and sequenced,
+can be used as a proxy for assessing changes in clonal abundance and better understanding population dynamics.
+`Pycashier` was originally developed for use with the
+ClonMapper Barcoding System [@alkhafaji2018;@gardner2024;@gardner2022]),
+which is comprised of random 20 nucleotide barcodes integrated as both a functional gRNA and expressed transcript.
 `Pycashier` has been previously utilized to interrogate
 tumor heterogeneity in barcoded cancer cell-line models [@gutierrez2021;@johnson2020],
-however, it is generalizable to similar DNA barcoding systems with known flanking regions.
-
+however, it is generalizable to similar DNA barcoding systems with known flanking regions and expected length.
 
 # Statement of need
 
-DNA sequencing and DNA barcoding specifically, have become
+DNA sequencing and cellular DNA barcoding specifically, have become
 more common as a modality for the characterization of clonal and lineage-specific subpopulations of cells.
 As researchers leverage these technologies, they'll require tools easy to setup
 and use to facilitate downstream biological analysis.
 DNA barcode sequencing suffers from several sources of noise that must be accounted for prior to statistical analysis.
 This noise can arise in typical Polymerase Chain Reaction (PCR) preparation [@kebschull2015;@potapov2017] or during read-out [@manley2016].
-Historically, the analysis of DNA barcoding has relied on tailored computational workflows,
+Historically, the analysis of cellular DNA barcoding has relied on tailored computational workflows,
 such as TimeMachine [@emert2021], that are difficult to parameterize
-or extend to similarly designed DNA barcoding systems.
+or extend to similarly designed cellular DNA barcoding systems.
 Recently, there has been the development of several NextFlow-based techniques,
 such as BARtab [@holze2024] and NextClone [@putri2023],
-which offer improved support for generalized approaches to processing barcode sequencing data.
-However, they require familiarity with NextFlow, which may be uncommon for experimentalists.
-`Pycashier` aims to be simple to install and generalizable enough to be useful to the broader community
-while also providing a user friendly interface.
+In particular BARtab and it's associated post-processing library bartools, offer an end-to-end toolkit for barcoding analysis.
+As a more feature complete end-to-end toolkit BARtab differs from `pycashier` by including support for spatial transcriptomics data and reference-based processing of barcodes.Additionally, being based on NextFlow offers some advantages to these tools including sample-level parameterization and tighter control on system resources.
+However, experience using NextFlow may be uncommon for experimentalists.
+`Pycashier` aims to be simple to install and generalizable enough to be useful to the broader community while also providing a user friendly interface.
 
 # Implementation and Usage
 
 `Pycashier` was intentionally designed to be simple-to-use for both computational and experimental biologists.
 It accomplishes this by leveraging purpose-built software for an opinionated DNA barcode processing pipeline.
 `Pycashier` has a command-line interface (CLI) implemented in `python`.
[…]
