### 2024-09-11 brocklab/pycashier#29 @daylinmorgan
the cli reference page should have plaintext versions
Right now https://docs.brocklab.com/pycashier/cli.html has an `svg` for each of the `pycashier` subcommands to demonstrate the available flags.

It would be helpful to also include a plain text version of these so that it is easier for users to copy and paste.

It might be possible to include both by using [`sphinx-tabs`](https://sphinx-tabs.readthedocs.io/en/latest/)
