### 2024-09-11 brocklab/pycashier#30 @daylinmorgan
example usage of `pycashier`
The documentation is missing an example for how to use `pycashier extract`, and the expected outputs that users can follow along with.
