### 2024-09-11 brocklab/pycashier#31 @daylinmorgan
allow overriding hardcoded executables
Currently all the executables are hardcoded as plain strings, meaning if someone has installed a tool in a non-standard location or under a different name it's not going to be used by `pycashier`.

Provide an option or environment variable to override this.
