### 2024-09-11 #6891 @daylinmorgan
Thank you both for the helpful suggestions and questions! I've pinned brocklab/pycashier#32 to track any issues associated with review comments.

> Authors should also consider providing some worked examples that a prospective user can run on their machines

> In the same idea, an example usage is required by JOSS, at least in the documentation. I didn’t find one.

As you've both highlighted there is need for a clear example use case of the technology that users can follow along with. I will add this to the documentation brocklab/pycashier#30.

---

@LTLA,

I've expanded the text of the paper to better describe the supported barcode configuration, of an expected length sequence with known flanking regions. These flanking regions are defined by the `-ua/--upstream-adapter` and `-da/--downstream-adapter` flags.

I've created brocklab/pycashier#29 to track providing plain text examples of the CLI flags.

While I think it's possible for users to customize the called utilities (starcode, cutadapt) by modifying their position in the `PATH`, I appreciate this isn't the most straightforward user experience and will add the ability to override these brocklab/pycashier#31.

---

@adrientaudiere,

I've expanded the summary to try to better clarify what is meant by DNA barcoding and the types of studies that it lends to. Additionally, I've elaborated more within the statement of need how `pycashier` compares to `BarTab` and `NextClone`. I've also added a schematic to the paper to help illustrate the associated inputs, commands and outputs.

> Using a personal fastq.files I obtained an error "failed to convert fastq to tsv"...I wonder what are the special structures of fastq files needed for Pycashier.

There isn't any special structure required so I'm not sure why it failed without more information about your inputs fastq. One possibility is that since `pycashier` will attempt to use existing files, if a previous invocation failed it might have tried to use a possibly empty output file. For example if the `sample.q30.barcode.fastq` is empty from `cutadapt` it will fail to convert an empty fastq to a tsv.

Regarding the `--threads` flag, Holze et al. are correct, this is simply passed onto underlying tools. `Pycashier` itself processes each sample serially.
