### 2024-09-12 132f5c3
fix: make receipt output deterministic
