### 2024-09-12 3c642a5
fix: use pixi to make docs
