### 2024-09-12 2add872
fix: use pixi to make docs
