### 2024-09-12 ec0cae9
chore: use pixi in makefile
