+++ pyproject.toml
@@ -67,7 +67,7 @@ platforms = ["linux-64"]
 
 
 [tool.pixi.environments]
-dev = ["base", "bioconda", "dev", "build", "docs", "tutorial"]
+dev = ["base", "bioconda", "dev", "build", "docs", "tutorial", "test"]
 docs = ["base", "docs", "tutorial"]
 prod = ["base", "bioconda", "prod"]
 build = ["base", "build"]
