### 2024-09-17 89c9197
chore: add deps to tutorial
