### 2024-09-19 54b0aa2
ci: add tests for python 3.11 and 3.12
