### 2024-09-26 #6891 @lpantano
Hi @LTLA and @adrientaudiere, are those comments helpful to move forward with next comments if any? Thanks!
