### 2024-09-26 4f7821d
fix(#31): add support for overriding runtime dependencies
