+++ src/pycashier/deps.py
@@ -1,0 +1,5 @@
+import os
+
+cutadapt = os.getenv("PYCASHIER_CUTADAPT", "cutadapt")
+fastp = os.getenv("PYCASHIER_FASTP", "fastp")
+starcode = os.getenv("PYCASHIER_STARCODE", "starcode")
