### 2024-09-26 ec1eb8c
fix(#31): add support for overriding runtime dependencies
