### 2024-09-26 1d8a694
fix(#28): remove erroneous print
