### 2024-09-26 6df1ecd
feat: improve sample display
