### 2024-09-27 45a8048
docs: switch to sphinx-inline-tabs
