+++ pyproject.toml
@@ -80,12 +80,12 @@ dev = [
 docs = ["base", "pycashier", "docs", "tutorial"]
 prod = ["base", "bioconda", "prod"]
 build = ["base", "build"]
-test = ["base", "bioconda", "dev","test", "pycashier"]
-test38 = ["base", "bioconda", "dev", "test", "pycashier", "test38"]
-test39 = ["base", "bioconda", "dev", "test", "pycashier", "test39"]
-test310 = ["base", "bioconda", "dev", "test", "pycashier", "test310"]
-test311 = ["base", "bioconda", "dev", "test", "pycashier", "test311"]
-test312 = ["base", "bioconda", "dev", "test", "pycashier", "test312"]
+test = ["base", "bioconda", "dev", "test", "pycashier", "prod"]
+test38 = ["base", "bioconda", "dev", "test", "pycashier", "prod", "test38"]
+test39 = ["base", "bioconda", "dev", "test", "pycashier", "prod", "test39"]
+test310 = ["base", "bioconda", "dev", "test", "pycashier", "prod", "test310"]
+test311 = ["base", "bioconda", "dev", "test", "pycashier", "prod", "test311"]
+test312 = ["base", "bioconda", "dev", "test", "pycashier", "prod", "test312"]
 
 
 [tool.pixi.host-dependencies]
@@ -104,6 +104,7 @@ sphinx-copybutton = "*"
 sphinx-inline-tabs = "*"
 myst-parser = "*"
 shibuya = "*"
+
 
 [tool.pixi.feature.pycashier.pypi-dependencies]
 pycashier = { path = ".", editable = true }
