### 2024-09-30 #6891 @lpantano
@daylinmorgan , can you go over the first items of the list above? Thanks!
