### 2024-09-30 b463219
fix: add back newline in checks
