### 2024-09-30 6c444bf
chore: add newlines to contributing.md
