+++ CHANGELOG.md
@@ -44,15 +44,18 @@ i.e. 1001, 1002, ..., 1999, 22000
 ## [2024.1003] - 2024-03-21
 
 ### Changed
+
 - Docker container uses pixi for tighter control (actual lock file)
 
 ### Fixed
+
 - warn user if there is a permission error on local directory
 - #22
 
 ## [2024.1002] - 2024-02-22
 
 ### Fixed
+
 - #21
 
 ## [2024.1001] - 2024-02-21
@@ -61,6 +64,7 @@ This release followed a major refactor and some things may be missing from the c
 Please see the documentation for up to date usage instructions.
 
 ### Added
+
 - sphinx-based documentation website
 - a log file produced in `./pipeline` which includes all subproccess output
 - added polars to dependencies to improve tabulated data operations
@@ -68,6 +72,7 @@ Please see the documentation for up to date usage instructions.
 - hidden `checks` command to invoke pre-run checks on demand for debugging
 
 ### Changed
+
 - new version scheme of YYYY.BUILD
 - pycashier will proceed even if a sample fails
 - output is less verbose
@@ -75,105 +80,141 @@ Please see the documentation for up to date usage instructions.
 - headers are included in all final tsvs
 
 ### Fixed
+
 - visual indicator of work when running counts on files (#18)
 - filename regex for merge is less stringent (#17)
 - won't fail if unneeded program is missing (#16)
 
 ## [23.1.2] - 2023-01-07
+
 ### Changed
+
 - updated ci pipeline to incorporate test prior to a tagged release
 
 ### Fixed
+
 - remove `|` type operator for python <3.10 compatibility (#13)
 - ensure init-check is run (#14)
 
 ## [23.1.1] - 2023-01-05
+
 ### Added
+
 - sample parameter for all subcommands
 - added some broad input/output tests
 - global parameters (i.e. threads or samples to fallback on)
+
 ### Changed
+
 - update python version requirements to >=3.8,<3.11
 - updated conda lock file for docker and pinned based images to sha256 for reproducibility
+
 ### Fixed
+
 - stop merge if both R1 and R2 not found
 - removed extra whitespace in single cell output tsv
 - added warning about broken symlinks (especially for docker) (#12)
 
 ## [22.10.1] - 2022-10-21
+
 ### Added
+
 - typechecking w/Mypy
+
 ### Changed
+
 - pinned docker image base to stable tag
 
 ## [22.9.1] - 2022-09-11
+
 ### Changed
+
 - made init check dependent on command run (fixes #10)
 - ignore hidden files in input directory (fixes #9)
 - simplified detection of PE reads
 - several UI changes for clarity and simplicity
 
 ### Added
+
 - check for config file and exit if it doesn't exist
 - separate `rich.console.Console` for stderr outputs
 - automated docker build and switched to ghcr.io
 
 ## [22.6.2] - 2022-06-30
+
 ### Added
+
 - new `-y/--yes` flag to skip prompts
 
 ### Changed
+
 - switch project management to pdm
 - improved merge error message
 - made docker build multi-stage to reduce footprint
 - updated ci to use python action instead of poetry
 
 ## [22.6.1] - 2022-06-27
+
 ### Fixed
+
 - Fixed(#8)
 
 ### Changed
+
 - refactored cli source for simplicity
 - project now adheres to [CalVer](https://calver.org)
 
 ## [0.3.5] - 2022-06-17
+
 ### Added
+
 - Encourage user to increase thread count
 
 ### Changed
+
 - Drop `just` for `make` to streamline development
 
 ## [0.3.4] - 2022-06-07
+
 ### Added
+
[…]
