### 2024-09-30 e1cb328
ci: use proper version for GHA step
