### 2024-09-30 #6891 @daylinmorgan
Hello @lpantano,

I've doubled checked all the authors and identifiers.

I've release a new version, `v2024.1006`, and uploaded an archive of this version to [zenodo](https://zenodo.org/records/13863889) with the DOI: https://doi.org/10.5281/zenodo.13863889.
