### 2024-10-01 #6891 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
