### 2024-10-05 71ae32f
paper: add missing citations
