+++ paper/paper.bib
@@ -13,6 +13,26 @@
   abstract = {Lineage tracking delivers essential quantitative insight into dynamic, probabilistic cellular processes, such as somatic tumor evolution and differentiation. Methods for high diversity lineage quantitation rely on sequencing a population of DNA barcodes. However, manipulation of specific individual lineages is not possible with this approach. To address this challenge, we developed a functionalized lineage tracing tool,~Control of Lineages~by Barcode Enabled Recombinant Transcription (COLBERT),~that enables high diversity lineage tracing and lineage-specific manipulation of gene expression. This modular platform utilizes expressed barcode gRNAs to both track cell lineages and direct lineage-specific gene expression.}
 }
 
+@article{bhang2015,
+  ids = {bhang2015a},
+  title = {Studying Clonal Dynamics in Response to Cancer Therapy Using High-Complexity Barcoding},
+  author = {Bhang, Hyo-eun C. and Ruddy, David A. and Krishnamurthy Radhakrishna, Viveksagar and Caushi, Justina X. and Zhao, Rui and Hims, Matthew M. and Singh, Angad P. and Kao, Iris and Rakiec, Daniel and Shaw, Pamela and Balak, Marissa and Raza, Alina and Ackley, Elizabeth and Keen, Nicholas and Schlabach, Michael R. and Palmer, Michael and Leary, Rebecca J. and Chiang, Derek Y. and Sellers, William R. and Michor, Franziska and Cooke, Vesselina G. and Korn, Joshua M. and Stegmeier, Frank},
+  year = {2015},
+  month = may,
+  journal = {Nature Medicine},
+  volume = {21},
+  number = {5},
+  pages = {440--448},
+  publisher = {Nature Publishing Group},
+  issn = {1546-170X},
+  doi = {10.1038/nm.3841},
+  abstract = {The authors have developed a barcoded library that enables the high-throughput tracking of tumor cell dynamics and clonal evolution in response to therapy.},
+  copyright = {2015 Nature Publishing Group, a division of Macmillan Publishers Limited. All Rights Reserved.},
+  langid = {english},
+  pmid = {25849130},
+  keywords = {Cancer therapeutic resistance,Genetic techniques,Targeted therapies,Tumour heterogeneity}
+}
+
 @article{chen2018,
   title = {Fastp: An Ultra-Fast All-in-One {{FASTQ}} Preprocessor},
   shorttitle = {Fastp},
@@ -50,17 +70,17 @@
 
 @incollection{gardner2022,
   title = {Functionalized {{Lineage Tracing}} for the {{Study}} and {{Manipulation}} of {{Heterogeneous Cell Populations}}},
-  booktitle = {Biomedical {{Engineering Technologies}}},
+  booktitle = {Biomedical {{Engineering Technologies}}: {{Volume}} 2},
   author = {Gardner, Andrea and Morgan, Daylin and Al'Khafaji, Aziz and Brock, Amy},
   editor = {Rasooly, Avraham and Baker, Houston and Ossandon, Miguel R.},
   year = {2022},
-  volume = {2394},
   pages = {109--131},
   publisher = {Springer US},
   address = {New York, NY},
   doi = {10.1007/978-1-0716-1811-0_8},
-  urldate = {2024-04-04},
-  isbn = {978-1-07-161810-3 978-1-07-161811-0},
+  urldate = {2024-10-04},
[…]
