### 2024-10-05 #6891 @daylinmorgan
When checking the proof there was a small grammatical error and missing citations. I've updated the manuscript and pushed the changes to the `joss` branch.
