### 2024-10-20 #6891 @Kevin-Mattheus-Moerman
@lpantano we cannot proceed at the moment since both reviewers should complete their checklists. Please work with the reviewers to get this completed and ping me when this is done.

@adrientaudiere, I see you are in principle happy with this work to be published in JOSS, however, I see some of your boxes are unticked. Can you please tick those boxes :point_up: or let us know if there are any open issues?

@LTLA I also see you said "LGTM", however we do unfortunately need the checklist completed to formally show the review was completed. Can you please call the following in a comment to generate the checklist:
`@editorialbot generate my checklist`
Next please tick all boxes that can be ticked. If you find any that cannot be ticked please alert the authors about what additional actions are needed.
