### 2024-10-29 f1afad8
chore: use pixi v0.34
