+++ .github/workflows/tests.yml
@@ -27,7 +27,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: prefix-dev/setup-pixi@v0.8.1
         with:
-          pixi-version: v0.28.2
+          pixi-version: v0.34.0
           cache: true
           manifest-path: pyproject.toml
           environments: ${{ matrix.environment }}
