+++ CHANGELOG.md
@@ -8,9 +8,9 @@ version -> vYYYY.BUILD
 where BUILD is a lexicographically increasing number:
 i.e. 1001, 1002, ..., 1999, 22000
 
-## [Unreleased]
+## Unreleased
 
-## [2024.1006] - 2024-09-30
+## 2024.1006 - 2024-09-30
 
 ### Changed
 
@@ -25,16 +25,16 @@ i.e. 1001, 1002, ..., 1999, 22000
 - proper exit code when nothing to process (#37)
 - remove erroneous debug print (#28)
 
-## [2024.1005] - 2024-09-19
+## 2024.1005 - 2024-09-19
 
 ### Changed
 
 -  change python reqs to >=3.8
 -  upgraded pixi to v0.28.2
 
-## [2024.1004] - 2024-04-02
+## 2024.1004 - 2024-04-02
 
-### Changed 
+### Changed
 
 - upgraded pixi to v0.18 for pyproject.toml support
 
@@ -43,7 +43,7 @@ i.e. 1001, 1002, ..., 1999, 22000
 - fix(#25)
 - actually fix(#17)
 
-## [2024.1003] - 2024-03-21
+## 2024.1003 - 2024-03-21
 
 ### Changed
 
@@ -54,13 +54,13 @@ i.e. 1001, 1002, ..., 1999, 22000
 - warn user if there is a permission error on local directory
 - #22
 
-## [2024.1002] - 2024-02-22
+## 2024.1002 - 2024-02-22
 
 ### Fixed
 
 - #21
 
-## [2024.1001] - 2024-02-21
+## 2024.1001 - 2024-02-21
 
 This release followed a major refactor and some things may be missing from the changelog.
 Please see the documentation for up to date usage instructions.
@@ -87,7 +87,7 @@ Please see the documentation for up to date usage instructions.
 - filename regex for merge is less stringent (#17)
 - won't fail if unneeded program is missing (#16)
 
-## [23.1.2] - 2023-01-07
+## 23.1.2 - 2023-01-07
 
 ### Changed
 
@@ -98,7 +98,7 @@ Please see the documentation for up to date usage instructions.
 - remove `|` type operator for python <3.10 compatibility (#13)
 - ensure init-check is run (#14)
 
-## [23.1.1] - 2023-01-05
+## 23.1.1 - 2023-01-05
 
 ### Added
 
@@ -117,7 +117,7 @@ Please see the documentation for up to date usage instructions.
 - removed extra whitespace in single cell output tsv
 - added warning about broken symlinks (especially for docker) (#12)
 
-## [22.10.1] - 2022-10-21
+## 22.10.1 - 2022-10-21
 
 ### Added
 
@@ -127,7 +127,7 @@ Please see the documentation for up to date usage instructions.
 
 - pinned docker image base to stable tag
 
-## [22.9.1] - 2022-09-11
+## 22.9.1 - 2022-09-11
 
 ### Changed
 
@@ -142,7 +142,7 @@ Please see the documentation for up to date usage instructions.
 - separate `rich.console.Console` for stderr outputs
 - automated docker build and switched to ghcr.io
 
-## [22.6.2] - 2022-06-30
+## 22.6.2 - 2022-06-30
 
 ### Added
 
@@ -155,7 +155,7 @@ Please see the documentation for up to date usage instructions.
 - made docker build multi-stage to reduce footprint
 - updated ci to use python action instead of poetry
 
-## [22.6.1] - 2022-06-27
+## 22.6.1 - 2022-06-27
 
 ### Fixed
 
@@ -166,7 +166,7 @@ Please see the documentation for up to date usage instructions.
[…]
