### 2024-10-29 425c0e5
build: bump polars minimum version to 1.10
