+++ pyproject.toml
@@ -7,7 +7,7 @@ dependencies = [
   "tomlkit>=0.10.1",
   "click-rich-help>=22.1.0",
   "click>=8.1.0",
-  "polars>=1.0.0",
+  "polars>=1.10.0",
 ]
 requires-python = ">=3.9"
 readme = "README.md"
@@ -94,7 +94,7 @@ pip = "*"
 python = ">=3.9,<3.13" # https://bioconda.github.io/faqs.html#what-versions-are-supported
 # when pixi add polars to pypi-dependencies it was trying to build it with cargo
 # adding it here to use pre-built objects from conda-forge
-polars = ">=1.0.0"
+polars = ">=1.10.0"
 
 [tool.pixi.feature.docs.dependencies]
 sphinx = "*"
@@ -117,7 +117,7 @@ rich = ">=12.0.0"
 tomlkit = ">=0.10.1"
 click-rich-help = ">=22.1.0"
 click = ">=8.1.0"
-polars = ">=1.0.0"
+polars = ">=1.10.0"
 
 [tool.pixi.feature.bioconda.dependencies]
 cutadapt = ">=4.6,<5"
