### 2024-10-29 fa87a63
fix(#42): allow ratio to be a float
