### 2024-10-31 #6891 @Kevin-Mattheus-Moerman
@daylinmorgan as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:
