### 2024-10-31 37d5316
docs: add citation and joss shield
