+++ CITATION.cff
@@ -1,0 +1,37 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Morgan
+  given-names: Daylin
+  orcid: "https://orcid.org/0000-0002-4218-2805"
+- family-names: Brock
+  given-names: Amy
+  orcid: "https://orcid.org/0000-0001-8255-9024"
+contact:
+- family-names: Brock
+  given-names: Amy
+  orcid: "https://orcid.org/0000-0001-8255-9024"
+doi: 10.5281/zenodo.13863889
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Morgan
+    given-names: Daylin
+    orcid: "https://orcid.org/0000-0002-4218-2805"
+  - family-names: Brock
+    given-names: Amy
+    orcid: "https://orcid.org/0000-0001-8255-9024"
+  date-published: 2024-10-31
+  doi: 10.21105/joss.06891
+  issn: 2475-9066
+  issue: 102
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 6891
+  title: "Pycashier: cash in on DNA barcode tags"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.06891"
+  volume: 9
+title: "Pycashier: cash in on DNA barcode tags"
+
