### 2024-10-31 1ced13f
docs: add more prominent link to paper
